Add files via upload
This is a super basic first draft, containing information from a table I created in postgresql. The coding syntax wasn't that difficult at all honestly it was properly installing and enabling php-pgsql on my computer. Most of the guides were for Windows file system not ubuntu (file extensions were different).